Renamed example package to amalitech
