+++ app/src/main/java/com/amalitech/bookmeetingroom/BookMeetingApplication.kt
@@ -15,6 +15,8 @@ import com.amalitech.onboarding_data.di.onboardingDataModule
 import com.amalitech.rooms.di.dataRoomModule
 import com.amalitech.rooms.di.domainRoomModule
 import com.amalitech.rooms.di.roomPresentationModule
+import com.amalitech.room.di.domainRoomModule
+import com.example.room.di.uiRoomModule
 import org.koin.android.ext.koin.androidContext
 import org.koin.core.context.GlobalContext.startKoin
