+++ domain/room/src/main/java/com/amalitech/room/book_room/model/Room.kt
@@ -1,0 +1,11 @@
+package com.amalitech.room.book_room.model
+
+import com.amalitech.core.domain.model.Booking
+
+data class Room(
+    val name: String,
+    val description: String,
+    val features: List<String>,
+    val bookings: List<Booking>,
+    val imgUrl: String
+)
