--- domain/room/src/main/java/com/example/room/book_room/use_case/BookRoom.kt
@@ -1,12 +1,0 @@
-package com.example.room.book_room.use_case
-
-import com.amalitech.core.domain.model.Booking
-import com.amalitech.core.util.UiText
-import kotlinx.coroutines.delay
-
-class BookRoom {
-    suspend operator fun invoke(booking: Booking): UiText? {
-        delay(2000)
-        return null
-    }
-}
