+++ ui/room/src/main/java/com/amalitech/room/book_room/SlotSelectionManager.kt
@@ -1,0 +1,8 @@
+package com.amalitech.room.book_room
+
+data class SlotSelectionManager(
+    val availableStartTimes: List<TimeUi> = emptyList(),
+    val availableEndTimes: List<TimeUi> = emptyList(),
+    val canShowStartTimes: Boolean = false,
+    val canShowEndTimes: Boolean = false
+)
