+++ ui/room/src/main/java/com/amalitech/room/book_room/TimeUi.kt
@@ -1,0 +1,8 @@
+package com.amalitech.room.book_room
+
+import java.time.LocalTime
+
+data class TimeUi(
+    val time: LocalTime,
+    val isAvailable: Boolean
+)
