+++ ui/room/src/main/java/com/amalitech/room/book_room/util/NavArguments.kt
@@ -1,0 +1,5 @@
+package com.amalitech.room.book_room.util
+
+object NavArguments {
+    const val roomId = "roomId"
+}
