+++ ui/room/src/main/java/com/amalitech/room/di/uiRoomModule.kt
@@ -1,0 +1,11 @@
+package com.amalitech.room.di
+
+import com.example.room.book_room.BookRoomViewModel
+import org.koin.androidx.viewmodel.dsl.viewModel
+import org.koin.dsl.module
+
+val uiRoomModule = module {
+    viewModel {
+        BookRoomViewModel(get())
+    }
+}
