--- ui/room/src/main/java/com/example/room/book_room/SlotSelectionManager.kt
@@ -1,8 +1,0 @@
-package com.example.room.book_room
-
-data class SlotSelectionManager(
-    val availableStartTimes: List<TimeUi> = emptyList(),
-    val availableEndTimes: List<TimeUi> = emptyList(),
-    val canShowStartTimes: Boolean = false,
-    val canShowEndTimes: Boolean = false
-)
