--- ui/room/src/main/java/com/example/room/book_room/TimeUi.kt
@@ -1,8 +1,0 @@
-package com.example.room.book_room
-
-import java.time.LocalTime
-
-data class TimeUi(
-    val time: LocalTime,
-    val isAvailable: Boolean
-)
