--- ui/room/src/main/java/com/example/room/book_room/components/BookRoomTitle.kt
@@ -1,23 +1,0 @@
-package com.example.room.book_room.components
-
-import androidx.compose.material3.MaterialTheme
-import androidx.compose.material3.Text
-import androidx.compose.runtime.Composable
-import androidx.compose.ui.Modifier
-import androidx.compose.ui.graphics.Color
-import androidx.compose.ui.text.TextStyle
-
-@Composable
-fun BookRoomTitle(
-    text: String,
-    modifier: Modifier = Modifier,
-    color: Color = MaterialTheme.colorScheme.onBackground,
-    textStyle: TextStyle = MaterialTheme.typography.titleLarge,
-) {
-    Text(
-        text = text,
-        style = textStyle,
-        modifier = modifier,
-        color = color
-    )
-}
