--- ui/room/src/main/java/com/example/room/book_room/util/NavArguments.kt
@@ -1,5 +1,0 @@
-package com.example.room.book_room.util
-
-object NavArguments {
-    const val roomId = "roomId"
-}
